+++ frontend/src/components/buttons/copy-to-clipboard-button.tsx
