+++ frontend/src/components/buttons/remove-button.tsx
