+++ frontend/src/components/buttons/stop-button.tsx
