+++ frontend/src/components/feedback-actions.tsx
@@ -1,6 +1,6 @@
 import ThumbsUpIcon from "#/icons/thumbs-up.svg?react";
 import ThumbDownIcon from "#/icons/thumbs-down.svg?react";
-import { FeedbackActionButton } from "./feedback-action-button";
+import { FeedbackActionButton } from "./buttons/feedback-action-button";
 
 interface FeedbackActionsProps {
   onPositiveFeedback: () => void;
