+++ frontend/src/components/github-repositories-suggestion-box.tsx
@@ -4,7 +4,7 @@ import { SuggestionBox } from "#/routes/_oh._index/suggestion-box";
 import { ConnectToGitHubModal } from "./modals/connect-to-github-modal";
 import { ModalBackdrop } from "./modals/modal-backdrop";
 import { GitHubRepositorySelector } from "#/routes/_oh._index/github-repo-selector";
-import { ModalButton } from "./modal-button";
+import { ModalButton } from "./buttons/modal-button";
 import GitHubLogo from "#/assets/branding/github-logo.svg?react";
 
 interface GitHubRepositoriesSuggestionBoxProps {
