+++ frontend/src/components/modals/confirmation-modals/base-modal.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ModalButton } from "#/components/modal-button";
+import { ModalButton } from "#/components/buttons/modal-button";
 import { ModalBody } from "../modal-body";
 
 interface ButtonConfig {
