+++ frontend/src/components/modals/connect-to-github-modal.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from "react-i18next";
 import { ModalBody } from "./modal-body";
 import { CustomInput } from "../custom-input";
-import { ModalButton } from "../modal-button";
+import { ModalButton } from "../buttons/modal-button";
 import {
   BaseModalDescription,
   BaseModalTitle,
