+++ frontend/src/routes/_oh.app._index/file-explorer/buttons/refresh-icon-button.tsx
@@ -1,5 +1,5 @@
 import { IoIosRefresh } from "react-icons/io";
-import { IconButton } from "#/components/icon-button";
+import { IconButton } from "#/components/buttons/icon-button";
 
 interface RefreshIconButtonProps {
   onClick: () => void;
