+++ frontend/src/routes/_oh.app/chat-interface.tsx
@@ -12,8 +12,8 @@ import { generateAgentStateChangeEvent } from "#/services/agent-state-service";
 import { FeedbackModal } from "../../components/feedback-modal";
 import { useScrollToBottom } from "#/hooks/use-scroll-to-bottom";
 import { TypingIndicator } from "../../components/typing-indicator";
-import { ContinueButton } from "../../components/continue-button";
-import { ScrollToBottomButton } from "../../components/scroll-to-bottom-button";
+import { ContinueButton } from "../../components/buttons/continue-button";
+import { ScrollToBottomButton } from "../../components/buttons/scroll-to-bottom-button";
 import { useWsClient } from "#/context/ws-client-provider";
 import { Messages } from "./messages";
 import { LoadingSpinner } from "./loading-spinner";
